+++ test/actions.js
@@ -434,7 +434,7 @@ int main() {...}
 	});
 
 	test('ban reporter for local report',  async () => {
-		const reportsPage = await fetch('http://dev-jschan/test/manage/reports.html', {
+		const reportsPage = await fetch('http://localhost/test/manage/reports.html', {
 			headers: {
 				'cookie': sessionCookie,
 			},
@@ -477,7 +477,7 @@ int main() {...}
 	});
 
 	test('ban reporter for global report',  async () => {
-		const reportsPage = await fetch('http://dev-jschan/globalmanage/reports.html', {
+		const reportsPage = await fetch('http://localhost/globalmanage/reports.html', {
 			headers: {
 				'cookie': sessionCookie,
 			},